Make public functions take current buffer as parameter
move-to, rename-to and delete-buffer-and-file used to get the current
buffer by asking the API themselves. Now it is passed as an argument, so
the functions become are useful if called from other plugins as well.

To keep delete-buffer-and-file working even if called with a file not
loaded in a buffer, wipe-buffer now checks if the buffer exists before
attempting to wipe it, and renamed accordingly.